Complete header section of the layout
Used solely vanilla css; however, I plan on switching to using scss
for work on the nav menu and content sections. This is because I could
really use the nesting functionality to clean up my code and save time.

The biggest things I learned so far were to use outline: none !important
on inputs to get rid of the borders and allow more responsive desings.

I feel like I improved my understanding of how grid layouts work and
will be able to work much faster on future sections.

Fixes for the future would be to implement nesting and other scss
functionality and fix my input bar so that the cursor lines up with
the placeholder text

In total this section took me 2.5 hours