Configure Layout for Content Section
Currently working though a Bug with the window height being set too big
I think it has something to do with overflow of grid elements. Definitley
need to fix this.

After fixing the grid, adding a footer, and aligning menu items, I can
add the chat section and then fix the search bar. Then I will be done
with the moderator dashboard for the Hush website!